Improve accessibility and scaling
This commit includes colorblind related accessibility improvements.
The site can be better accessed through devices with smaller screens.
The add game form can be triggered by pressing a button on devices with
a smaller screen.

Bug: If the screen is resized to a "sidebar-compatible" width
while the form is displayed on a smaller screen, the sidebar will
not have a form.